split test cases
