+++ fava_investor/modules/performance/test_contributions.py
@@ -18,16 +18,6 @@ class TestContributions(SplitTestCase):
         self.assertEqual({}, result)
 
     @test_utils.docfile
-    def test_no_withdrawals(self, filename: str):
-        """
-        2020-01-01 open Assets:Bank
-        2020-01-01 open Assets:Account
-        2020-01-01 open Assets:Account:Sub
-        """
-        result = sum_inventories(get_split(filename).withdrawals)
-        self.assertEqual({}, result)
-
-    @test_utils.docfile
     def test_contributions_to_subaccounts(self, filename: str):
         """
         2020-01-01 open Assets:Bank
@@ -63,39 +53,6 @@ class TestContributions(SplitTestCase):
         """
         result = sum_inventories(get_split(filename).contributions)
         self.assertEqual({}, result)
-
-    @test_utils.docfile
-    def test_list_withdrawals_entries(self, filename: str):
-        """
-        2020-01-01 open Assets:Bank
-        2020-01-01 open Assets:Account:A
-        2020-01-01 open Assets:Account:B
-
-        2020-01-01 * "contrib"
-            Assets:Account:A  2 GBP
-            Assets:Bank
-
-        2020-01-02 * "withdrawal 1"
-            Assets:Account:A  -1 GBP
-            Assets:Account:B  -2 GBP
-            Assets:Bank
-
-        2020-01-02 * "withdrawal 2"
-            Assets:Account:A  -3 GBP
-            Assets:Bank
-        """
-        split = get_split(filename)
-
-        self.assertEqual({}, split.withdrawals[0])
-        self.assertEqual(i("-3 GBP"), split.withdrawals[1])
-        self.assertEqual(i("-3 GBP"), split.withdrawals[2])
-
-        self.assertIsInstance(split.transactions[0], Transaction)
-        self.assertIsInstance(split.transactions[1], Transaction)
-        self.assertIsInstance(split.transactions[2], Transaction)
-        self.assertEqual("contrib", split.transactions[0].narration)
-        self.assertEqual("withdrawal 1", split.transactions[1].narration)
-        self.assertEqual("withdrawal 2", split.transactions[2].narration)
 
     @test_utils.docfile
     def test_list_contribution_entries(self, filename: str):
@@ -142,7 +99,7 @@ class TestContributions(SplitTestCase):
             Assets:Account:B
         """
         self.skipTest(
-            "Not implemented. It will be needed to calculate returns for selected account as well."
+            "value account filtering not implemented"
         )
 
         split = get_split(filename)
@@ -167,19 +124,3 @@ class TestContributions(SplitTestCase):
         split = get_split(filename)
 
         self.assertEqual(Inventory.from_string("5 GBP"), sum_inventories(split.contributions))
-
-    @test_utils.docfile
-    def test_asset_sold_loan_returned_and_rest_withdrawn(self, filename: str):
-        """
-        2020-01-01 open Assets:Bank
-        2020-01-01 open Assets:Account:Loan
-        2020-01-01 open Assets:Account:Asset
-
-        2020-01-02 * "transfer"
-            Assets:Account:Loan  6 GBP
-            Assets:Account:Asset  -1 AA {11 GBP}
-            Assets:Bank  5 GBP
-        """
-        split = get_split(filename)
-
-        self.assertEqual(Inventory.from_string("-5 GBP"), sum_inventories(split.withdrawals))
