+++ fava_investor/modules/performance/test_dividends.py
@@ -54,29 +54,3 @@ class TestDividends(SplitTestCase):
         }
         split = get_split(filename, config)
         self.assertEqual(Inventory(), sum_inventories(split.dividends))
-
-    @test_utils.docfile
-    def test_cost(self, filename: str):
-        """
-        2020-01-01 open Assets:Account
-        2020-01-01 open Expenses:ServiceFee
-
-        2020-01-01 * "dividend"
-            Assets:Account  -5 GBP
-            Expenses:ServiceFee
-        """
-        split = get_split(filename)
-        inventories = sum_inventories(split.costs)
-        self.assertEqual(Inventory.from_string("-5 GBP"), inventories)
-
-    @test_utils.docfile
-    def test_sum_of_each_split_should_match_balance(self, filename):
-        """
-        2020-01-01 open Assets:Account
-        2020-01-01 open Expenses:ServiceFee
-        2020-01-02 * "cost"
-            Assets:Account  1 USD
-            Expenses:ServiceFee
-        """
-
-        self.assertSumOfSplitsEqualValue(filename)
