+++ fava_investor/modules/performance/test_gains.py
@@ -21,33 +21,6 @@ class TestGains(SplitTestCase):
         self.assertInventoriesSum("1 USD", split.gains_unrealized)
 
     @test_utils.docfile
-    def test_get_unrealized_gains_entries(self, filename: str):
-        """
-        2020-01-01 open Assets:Bank
-        2020-01-01 open Assets:Account:A
-        2020-01-01 open Assets:Account:B
-
-        2020-02-22 * "Buy stock"
-          Assets:Account:A  1 AA {1 USD}
-          Assets:Bank
-
-        2020-02-23 * "Buy stock"
-          Assets:Account:B  2 AA {1 USD}
-          Assets:Bank
-
-        2020-02-24 price AA  2 USD
-        """
-        self.skipTest("todo ")
-        sut = get_sut(filename)
-        result = sut.get_unrealized_gains_per_account()
-
-        expected = {
-            "Assets:Account:A": {"USD": 1},
-            "Assets:Account:B": {"USD": 2},
-        }
-        self.assertEqual(expected, result)
-
-    @test_utils.docfile
     def test_unrealized_ignoreNonValueAccounts(self, filename: str):
         """
         2020-01-01 open Assets:Bank
